script/pyproject.toml
[tool.poetry]
name = "script"
version = "0.1.01"
description = ""
authors = ["Takahisa Kurimura <kuri.226.t@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^3.141.0"
slackclient = "^2.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
